Update of DMP data function and added calibration
Rewrote and renamed the function that reads data from DMP, based on updated example code. Overflow no longer seems to be an issue.
Calibration routin was added in the setup(). Offsets printed to serial, still needs to be changed manually.
Removed old comments and restructured the code.
Changed GPIO from 10 to 9 for LedRing du to space issues in hardware.